Reorganize random_circuit in new file
